Theory: The File type signals the potential for filesystem operations
The PLINK file is actually a prefix, appended with .{bed,bim,fam} by
SAIGE. However, the pipeline runner tried to copy the prefix file
around, which obviously didn't work. I've changed its type to String and
explicitly demarcated the files to which it refers (so they can be
copied to where they need to be)... Fingers crossed!